+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Support for HTTPS cloning using Github username and token provided to atlantis server ([#117](https://github.com/hootsuite/atlantis/pull/117))
 * Set custom user agent for AWS interactions by atlantis server ([#115](https://github.com/hootsuite/atlantis/pull/115))
 * Adding `post_plan` and `post_apply` commands ([#102](https://github.com/hootsuite/atlantis/pull/102))
+* Adding the ability to verify webhook secret ([#120](https://github.com/hootsuite/atlantis/pull/120))
 
 ### Bug Fixes